+++ plugins/css/deps.ts
@@ -1,1 +1,1 @@
-export { parse } from "https://deno.land/std@0.87.0/path/mod.ts";
+export { parse } from "https://deno.land/std@0.91.0/path/mod.ts";