+++ plugins/importmap/deps.ts
@@ -3,4 +3,4 @@ export {
   fromFileUrl,
   normalize,
   relative,
-} from "https://deno.land/std@0.87.0/path/mod.ts";
+} from "https://deno.land/std@0.91.0/path/mod.ts";