+++ plugins/postcss/deps.ts
@@ -4,4 +4,4 @@ export type {
 } from "https://deno.land/x/postcss@8.2.6/lib/postcss.d.ts";
 export { default as postcss } from "https://deno.land/x/postcss@8.2.6/mod.js";
 export { default as postcssModules } from "https://esm.sh/postcss-modules@4.0.0";
-export { parse } from "https://deno.land/std@0.87.0/path/mod.ts";
+export { parse } from "https://deno.land/std@0.91.0/path/mod.ts";