+++ src/error.ts
@@ -1,12 +1,12 @@
 /**
- * Derived from <https://github.com/rollup/rollup/blob/v2.39.0/src/utils/error.ts>
+ * Derived from <https://github.com/rollup/rollup/blob/v2.42.3/src/utils/error.ts>
  */
 
 import type { RollupError } from "../deps.ts";
 
 /**
  * error
- * 
+ *
  * @param {Error|RollupError} base
  * @private
  */