+++ src/relativeId.ts
@@ -1,12 +1,12 @@
 /**
- * Derived from <https://github.com/rollup/rollup/blob/v2.39.0/src/utils/relativeId.ts>
+ * Derived from <https://github.com/rollup/rollup/blob/v2.42.3/src/utils/relativeId.ts>
  */
 
 import { isAbsolute, relative } from "../deps.ts";
 
 /**
  * relativeId
- * 
+ *
  * @param {string} id
  * @returns {string}
  * @private