+++ src/rollup/mod.ts
@@ -1,17 +1,10 @@
-// deno-lint-ignore-file ban-types
 import type {
-  ExternalOption,
-  InputOption,
-  ManualChunksOption,
+  InputOptions as _InputOptions,
   OutputOptions,
-  Plugin,
-  PreserveEntrySignaturesOption,
-  RollupCache,
   RollupWatcherEvent,
-  TreeshakingOptions,
-  WarningHandlerWithDefault,
 } from "../../deps.ts";
 import { VERSION } from "../../deps.ts";
+import { DenoResolverOptions } from "../rollup-plugin-deno-resolver/denoResolver.ts";
 import { rollup } from "./rollup.ts";
 import { watch } from "./watch.ts";
 
@@ -23,7 +16,7 @@ export { rollup, VERSION, watch };
 /**
  * Prevent `warning: Compiled module not found` by re-exporting
  * types explicitly.
- * 
+ *
  * @public
  */
 export type {
@@ -102,7 +95,6 @@ export type {
   RollupCache,
   RollupError,
   RollupLogProps,
-  RollupOptions,
   RollupOutput,
   RollupWarning,
   RollupWatcherEvent,
@@ -162,32 +154,16 @@ export interface RollupWatcher extends
 /**
  * @public
  */
-export interface InputOptions {
-  acorn?: Object;
-  acornInjectPlugins?: Function | Function[];
-  cache?: false | RollupCache;
-  context?: string;
-  experimentalCacheExpiry?: number;
-  external?: ExternalOption;
-  /** @deprecated Use the "inlineDynamicImports" output option instead. */
-  inlineDynamicImports?: boolean;
-  input?: InputOption;
-  /** @deprecated Use the "manualChunks" output option instead. */
-  manualChunks?: ManualChunksOption;
-  moduleContext?: ((id: string) => string | null | undefined) | {
-    [id: string]: string;
-  };
-  onwarn?: WarningHandlerWithDefault;
-  perf?: boolean;
-  plugins?: Plugin[];
-  preserveEntrySignatures?: PreserveEntrySignaturesOption;
-  /** @deprecated Use the "preserveModules" output option instead. */
-  preserveModules?: boolean;
-  preserveSymlinks?: boolean;
-  shimMissingExports?: boolean;
-  strictDeprecations?: boolean;
-  treeshake?: boolean | TreeshakingOptions;
-  watch?: WatcherOptions | false;
+export interface InputOptions extends _InputOptions {
+  denoResolver?: DenoResolverOptions;
+}
+
+/**
+ * @public
+ */
+export interface RollupOptions extends InputOptions {
+  // This is included for compatibility with config files but ignored by rollup.rollup
+  output?: OutputOptions | OutputOptions[];
 }
 
 /**