+++ test/deps.ts
@@ -1,9 +1,9 @@
-export * as posixPath from "https://deno.land/std@0.87.0/path/posix.ts";
-export { createRequire } from "https://deno.land/std@0.87.0/node/module.ts";
+export * as posixPath from "https://deno.land/std@0.97.0/path/posix.ts";
+export { createRequire } from "https://deno.land/std@0.97.0/node/module.ts";
 export { expect } from "https://deno.land/x/expect@v0.2.6/mod.ts";
 export { default as assert } from "https://esm.sh/assert@2.0.0";
 export { default as listGitHub } from "https://esm.sh/list-github-dir-content@3.0.0";
 export { default as magicString } from "https://esm.sh/magic-string@0.25.7";
 export { default as url } from "https://esm.sh/url@0.11.0";
 export { default as sourceMap } from "https://esm.sh/source-map@0.7.3";
-export { default as acorn } from "https://esm.sh/acorn@8.0.4";
+export { default as acorn } from "https://cdn.skypack.dev/acorn@8.0.4";